Calculate perimeter and area of data scatter

This function determines the perimeter and area around the boundary points

Usage

AP(dat)

Arguments

dat

A dataframe with two columns.

Value

The area and perimeter covered bivariate data points, and the selected boundary points

Examples

x<-data.frame(x=evapotranspiration$`ET(mm)`,y=evapotranspiration$`yield(t/ha)`)
AP(x)
#> $Perimeter
#> [1] 2758.392
#> 
#> $Area
#> [1] 858.116
#> 
#> $Polygon
#>            x     y selection
#> 418  108.619 0.800       max
#> 538  114.000 1.070       max
#> 533  116.000 1.640       max
#> 154  120.000 0.960       max
#> 155  123.750 0.990       max
#> 430  127.445 1.500       max
#> 421  129.827 1.400       max
#> 422  132.095 1.000       max
#> 535  133.000 1.450       max
#> 383  135.703 1.000       max
#> 451  137.150 1.900       max
#> 523  139.000 0.960       max
#> 382  139.074 1.100       max
#> 98   139.474 0.530       max
#> 534  141.000 2.090       max
#> 169  142.857 1.000       max
#> 536  144.000 1.660       max
#> 387  144.678 0.800       max
#> 522  146.000 1.250       max
#> 532  150.000 0.440       max
#> 537  151.000 1.590       max
#> 360  152.000 0.740       max
#> 230  153.000 1.390       max
#> 605  153.828 2.103       max
#> 521  155.000 1.330       max
#> 23   158.156 2.230       max
#> 168  158.333 0.950       max
#> 232  159.000 1.520       max
#> 134  160.000 0.320       max
#> 431  162.437 1.400       max
#> 231  164.000 1.680       max
#> 20   164.773 1.450       max
#> 99   166.071 0.930       max
#> 149  166.667 0.500       max
#> 77   167.000 1.554       max
#> 417  171.065 1.200       max
#> 484  173.000 2.600       max
#> 419  174.062 1.000       max
#> 171  174.286 1.220       max
#> 531  175.000 0.870       max
#> 86   175.472 0.930       max
#> 440  176.100 0.800       max
#> 170  177.500 1.420       max
#> 21   177.698 2.470       max
#> 514  178.000 2.010       max
#> 388  179.462 1.100       max
#> 79   180.000 1.887       max
#> 433  180.170 1.300       max
#> 87   181.046 2.770       max
#> 452  181.190 2.100       max
#> 648  182.858 2.342       max
#> 76   183.000 2.832       max
#> 403  184.232 1.100       max
#> 439  184.810 1.300       max
#> 496  185.000 0.600       max
#> 143  185.714 1.300       max
#> 450  185.720 1.900       max
#> 602  185.796 2.333       max
#> 530  186.000 0.940       max
#> 96   186.567 1.250       max
#> 142  186.667 1.120       max
#> 487  188.000 2.840       max
#> 399  188.440 1.900       max
#> 336  189.000 1.660       max
#> 447  190.390 1.600       max
#> 162  190.833 2.290       max
#> 163  191.000 1.910       max
#> 157  191.667 1.150       max
#> 435  192.310 1.800       max
#> 424  192.752 1.600       max
#> 410  194.031 1.600       max
#> 488  195.000 3.210       max
#> 436  195.660 2.300       max
#> 30   195.683 2.720       max
#> 444  195.700 2.600       max
#> 513  196.000 2.070       max
#> 233  197.000 0.170       max
#> 147  197.500 0.790       max
#> 242  199.000 0.180       max
#> 658  199.591 3.348       max
#> 486  200.000 3.070       max
#> 414  201.173 1.000       max
#> 574  201.344 2.028       max
#> 29   201.408 2.860       max
#> 278  202.000 0.890       max
#> 136  202.500 0.810       max
#> 611  202.923 1.794       max
#> 512  203.000 2.450       max
#> 541  204.000 3.300       max
#> 591  204.174 3.201       max
#> 603  204.835 3.169       max
#> 148  205.000 0.410       max
#> 606  205.466 1.728       max
#> 494  206.000 1.900       max
#> 607  206.334 1.851       max
#> 74   207.000 2.823       max
#> 397  207.195 1.900       max
#> 22   207.937 2.620       max
#> 428  209.027 2.100       max
#> 401  209.551 2.000       max
#> 380  210.000 1.050       max
#> 511  211.000 1.730       max
#> 449  211.390 2.200       max
#> 659  211.544 3.522       max
#> 590  212.128 3.580       max
#> 576  212.946 2.532       max
#> 10   213.000 1.750       max
#> 442  213.970 2.600       max
#> 539  214.000 1.950       max
#> 655  214.563 3.119       max
#> 572  214.733 2.035       max
#> 71   215.000 3.233       max
#> 94   215.152 0.710       max
#> 75   216.000 3.062       max
#> 643  216.162 1.441       max
#> 434  216.250 3.300       max
#> 589  216.316 2.881       max
#> 146  216.667 0.650       max
#> 650  217.229 1.375       max
#> 404  218.055 2.000       max
#> 646  218.428 1.304       max
#> 579  218.624 2.025       max
#> 578  218.788 2.810       max
#> 385  218.847 2.900       max
#> 97   218.947 2.080       max
#> 448  219.900 1.600       max
#> 78   220.000 2.750       max
#> 313  220.218 1.170       max
#> 28   220.339 2.600       max
#> 689  220.688 3.772       max
#> 652  221.026 3.072       max
#> 654  221.239 2.946       max
#> 651  221.884 2.107       max
#> 420  222.062 1.800       max
#> 292  222.862 1.433       max
#> 540  223.000 3.840       max
#> 588  223.029 2.537       max
#> 437  223.380 3.400       max
#> 642  223.715 2.169       max
#> 656  223.777 4.694       max
#> 580  223.794 1.813       max
#> 649  223.954 1.836       max
#> 432  225.087 0.000       max
#> 690  225.965 4.564       max
#> 331  226.000 1.500       max
#> 575  226.040 2.307       max
#> 660  226.430 2.585       max
#> 446  226.760 1.500       max
#> 443  226.990 2.900       max
#> 529  227.000 2.020       max
#> 95   227.027 0.840       max
#> 577  228.264 2.224       max
#> 423  228.401 0.800       max
#> 653  228.449 2.947       max
#> 600  228.925 2.202       max
#> 345  229.000 1.340       max
#> 166  229.167 2.750       max
#> 586  229.303 2.364       max
#> 627  229.636 2.963       max
#> 509  230.000 1.960       max
#> 93   230.435 1.060       max
#> 384  231.986 1.300       max
#> 682  232.768 4.013       max
#> 675  232.798 3.707       max
#> 502  233.000 1.550       max
#> 657  233.059 3.092       max
#> 585  233.474 3.139       max
#> 616  233.607 2.647       max
#> 526  234.000 2.560       max
#> 408  234.004 1.600       max
#> 151  235.000 0.470       max
#> 393  235.533 1.400       max
#> 594  235.606 2.543       max
#> 167  235.833 2.830       max
#> 510  236.000 2.320       max
#> 587  236.408 3.656       max
#> 583  236.518 3.012       max
#> 504  237.000 2.680       max
#> 508  238.000 2.310       max
#> 409  239.055 1.900       max
#> 438  239.620 3.800       max
#> 411  239.684 1.300       max
#> 492  240.000 2.130       max
#> 644  240.173 2.181       max
#> 619  240.594 2.778       max
#> 416  240.735 1.900       max
#> 622  240.809 3.895       max
#> 269  241.000 1.090       max
#> 527  242.000 2.250       max
#> 406  242.430 3.200       max
#> 197  243.000 2.760       max
#> 453  243.310 2.400       max
#> 679  243.530 2.929       max
#> 396  243.874 1.200       max
#> 528  244.000 2.340       max
#> 372  245.000 1.550       max
#> 663  245.713 3.502       max
#> 599  245.947 2.001       max
#> 525  246.000 2.810       max
#> 402  246.302 2.100       max
#> 464  247.000 1.060       max
#> 548  247.045 3.779       max
#> 582  247.713 2.403       max
#> 645  247.799 1.699       max
#> 601  249.135 2.050       max
#> 390  249.330 1.400       max
#> 427  249.968 4.700       max
#> 506  250.000 2.680       max
#> 598  250.231 2.056       max
#> 407  250.274 2.100       max
#> 604  250.559 2.391       max
#> 301  250.952 3.289       max
#> 610  251.011 1.597       max
#> 322  251.229 1.829       max
#> 687  251.300 5.614       max
#> 405  251.548 2.200       max
#> 542  251.704 2.858       max
#> 584  252.197 3.468       max
#> 386  252.373 0.800       max
#> 85   253.571 2.130       max
#> 429  253.921 1.900       max
#> 324  254.000 2.065       max
#> 31   254.783 2.930       max
#> 546  254.939 3.129       max
#> 73   255.000 3.537       max
#> 547  255.216 2.709       max
#> 381  255.546 0.600       max
#> 49   256.000 2.350       max
#> 323  256.532 2.044       max
#> 673  256.611 3.486       max
#> 621  256.744 2.832       max
#> 628  256.956 3.134       max
#> 612  257.721 2.540       max
#> 683  257.969 4.970       max
#> 235  258.000 1.260       max
#> 688  258.123 3.421       max
#> 641  258.874 2.122       max
#> 551  258.962 3.307       max
#> 474  260.000 2.160       max
#> 678  260.076 3.670       max
#> 389  260.487 1.700       max
#> 620  260.686 3.135       max
#> 595  260.873 1.880       max
#> 638  261.681 2.327       max
#> 490  262.000 2.400       max
#> 608  262.445 1.914       max
#> 195  263.000 2.400       max
#> 640  263.021 2.690       max
#> 680  263.293 2.631       max
#> 305  263.398 2.219       max
#> 544  263.581 2.763       max
#> 441  263.650 1.500       max
#> 573  263.937 2.229       max
#> 681  264.057 2.953       max
#> 549  264.751 3.030       max
#> 596  264.824 2.474       max
#> 613  265.293 2.045       max
#> 173  266.000 3.990       max
#> 391  266.355 2.000       max
#> 571  266.594 3.581       max
#> 593  266.631 2.777       max
#> 623  266.741 3.999       max
#> 9    267.000 2.493       max
#> 332  268.000 1.340       max
#> 196  269.000 2.820       max
#> 545  269.122 2.667       max
#> 550  269.607 3.110       max
#> 691  269.746 4.755       max
#> 413  270.302 2.100       max
#> 445  270.420 1.300       max
#> 92   270.833 0.650       max
#> 34   271.000 1.130       max
#> 592  271.708 2.358       max
#> 614  271.716 2.614       max
#> 635  273.153 1.636       max
#> 553  273.266 4.717       max
#> 617  273.375 2.779       max
#> 294  274.311 3.086       max
#> 618  274.652 2.461       max
#> 192  275.000 3.090       max
#> 172  275.714 3.860       max
#> 477  276.000 2.370       max
#> 543  276.871 3.973       max
#> 479  277.000 3.360       max
#> 609  277.125 2.450       max
#> 41   278.000 2.510       max
#> 661  278.293 2.771       max
#> 684  278.486 4.609       max
#> 415  279.160 1.400       max
#> 639  280.221 3.069       max
#> 581  280.543 2.015       max
#> 303  280.924 2.596       max
#> 33   281.000 2.580       max
#> 552  281.270 4.458       max
#> 668  281.378 2.891       max
#> 297  281.432 2.367       max
#> 561  281.540 4.470       max
#> 193  282.000 3.440       max
#> 293  282.015 2.603       max
#> 564  282.554 4.694       max
#> 632  282.924 3.963       max
#> 359  283.000 1.320       max
#> 676  283.209 2.598       max
#> 637  283.742 1.795       max
#> 45   284.000 2.270       max
#> 677  284.446 3.721       max
#> 287  284.555 1.824       max
#> 478  285.000 2.240       max
#> 554  285.943 4.852       max
#> 295  286.415 2.804       max
#> 320  286.738 3.416       max
#> 672  287.051 1.776       max
#> 319  287.341 3.188       max
#> 597  287.354 2.220       max
#> 296  287.660 2.914       max
#> 298  287.782 4.008       max
#> 425  287.973 3.400       max
#> 153  288.182 3.170       max
#> 566  288.525 4.203       max
#> 631  288.744 3.174       max
#> 476  289.000 2.500       max
#> 615  289.048 2.980       max
#> 647  289.700 3.131       max
#> 309  289.814 3.436       max
#> 140  290.000 1.160       max
#> 626  290.345 2.555       max
#> 563  290.752 3.794       max
#> 630  290.987 3.315       max
#> 194  291.000 3.380       max
#> 398  291.135 1.900       max
#> 519  292.000 3.510       max
#> 152  292.500 2.340       max
#> 565  292.992 6.065       max
#> 517  293.000 3.390       max
#> 556  293.307 4.791       max
#> 671  293.922 3.010       max
#> 315  294.386 3.517       max
#> 568  294.759 4.291       max
#> 187  295.000 2.350       max
#> 666  295.141 3.264       max
#> 189  296.000 3.290       max
#> 636  296.106 2.111       max
#> 280  297.000 1.270       max
#> 314  297.434 2.145       max
#> 674  297.950 3.272       max
#> 256  298.000 2.460       max
#> 299  299.212 2.818       max
#> 318  299.287 2.044       max
#> 203  300.000 2.570       max
#> 562  300.810 4.551       max
#> 665  300.940 3.077       max
#> 670  300.953 2.663       max
#> 288  300.987 2.134       max
#> 100  301.000 1.830       max
#> 685  301.220 5.806       max
#> 141  301.667 1.810       max
#> 629  302.059 2.991       max
#> 392  302.245 2.000       max
#> 321  302.249 3.665       max
#> 633  302.636 2.397       max
#> 498  303.000 3.200       max
#> 686  303.518 5.644       max
#> 338  304.000 3.240       max
#> 570  304.164 4.360       max
#> 569  304.555 5.234       max
#> 555  304.888 5.372       max
#> 481  305.000 3.580       max
#> 624  305.258 3.634       max
#> 634  305.991 1.919       max
#> 362  306.000 3.190       max
#> 559  306.502 4.519       max
#> 244  307.000 1.190       max
#> 662  307.093 3.242       max
#> 60   309.000 2.700       max
#> 145  310.000 4.960       max
#> 316  310.184 1.782       max
#> 104  311.000 1.759       max
#> 557  311.579 5.070       max
#> 395  311.892 1.600       max
#> 190  312.000 3.380       max
#> 667  312.140 2.589       max
#> 412  312.950 2.100       max
#> 483  313.000 3.680       max
#> 371  314.000 2.080       max
#> 274  315.000 2.510       max
#> 317  315.347 1.970       max
#> 133  316.000 3.224       max
#> 132  317.000 3.660       max
#> 184  320.000 3.320       max
#> 567  320.944 4.746       max
#> 265  322.000 1.930       max
#> 497  323.000 3.270       max
#> 426  323.353 2.100       max
#> 333  325.000 4.120       max
#> 368  326.000 2.550       max
#> 165  326.875 5.230       max
#> 337  327.000 2.750       max
#> 290  327.894 2.833       max
#> 473  328.000 4.690       max
#> 91   328.261 3.020       max
#> 201  329.000 2.090       max
#> 90   329.412 2.800       max
#> 300  329.946 2.872       max
#> 131  330.000 1.618       max
#> 289  330.713 2.692       max
#> 164  330.769 4.300       max
#> 178  331.000 3.380       max
#> 394  331.809 2.200       max
#> 126  333.000 1.966       max
#> 669  333.161 2.959       max
#> 291  333.844 2.397       max
#> 175  334.000 3.390       max
#> 179  335.000 4.030       max
#> 664  335.316 3.321       max
#> 81   336.000 2.822       max
#> 469  338.000 3.670       max
#> 176  339.000 3.700       max
#> 181  340.000 3.450       max
#> 174  341.000 3.640       max
#> 182  342.000 3.390       max
#> 204  343.000 1.460       max
#> 144  343.333 2.060       max
#> 558  343.679 4.513       max
#> 177  344.000 3.440       max
#> 625  344.806 4.338       max
#> 467  346.000 1.940       max
#> 180  347.000 3.440       max
#> 334  348.000 3.390       max
#> 302  348.488 3.087       max
#> 264  350.000 1.260       max
#> 459  351.000 3.010       max
#> 470  352.000 4.760       max
#> 312  352.044 3.289       max
#> 161  352.222 3.170       max
#> 109  354.000 2.433       max
#> 234  355.000 1.420       max
#> 307  355.854 3.638       max
#> 350  356.000 1.700       max
#> 158  356.250 2.850       max
#> 82   357.000 3.249       max
#> 108  358.000 2.225       max
#> 112  359.000 1.853       max
#> 12   361.000 1.603       max
#> 88   361.053 3.430       max
#> 458  362.000 3.890       max
#> 243  363.000 1.750       max
#> 16   364.474 2.770       max
#> 111  365.000 3.317       max
#> 114  366.000 2.656       max
#> 223  367.000 1.990       max
#> 228  368.000 0.380       max
#> 200  369.000 2.160       max
#> 400  369.658 2.300       max
#> 367  370.000 4.440       max
#> 465  371.000 2.650       max
#> 560  371.925 5.589       max
#> 277  372.000 1.540       max
#> 275  373.000 1.600       max
#> 89   374.194 3.480       max
#> 463  375.000 3.500       max
#> 304  375.412 3.833       max
#> 5    376.000 3.100       max
#> 222  377.000 2.110       max
#> 466  379.000 2.420       max
#> 186  380.000 5.600       max
#> 214  381.000 2.070       max
#> 327  381.984 2.609       max
#> 268  382.000 1.620       max
#> 18   384.536 3.730       max
#> 123  385.000 1.589       max
#> 471  386.000 3.870       max
#> 352  387.000 2.740       max
#> 462  388.000 3.380       max
#> 25   389.744 4.560       max
#> 1    390.000 3.270       max
#> 472  391.000 4.100       max
#> 17   391.176 3.990       max
#> 325  391.537 2.374       max
#> 121  392.000 1.373       max
#> 270  393.000 2.250       max
#> 364  395.000 2.600       max
#> 185  396.000 5.400       max
#> 32   398.000 2.030       max
#> 326  398.684 2.502       max
#> 13   399.000 2.542       max
#> 369  400.000 3.760       max
#> 24   400.952 4.210       max
#> 15   401.000 3.976       max
#> 261  402.000 2.180       max
#> 36   404.000 2.050       max
#> 26   404.040 4.000       max
#> 257  406.000 1.090       max
#> 215  407.000 2.270       max
#> 35   408.000 4.360       max
#> 19   409.524 4.300       max
#> 308  409.956 2.656       max
#> 273  410.000 1.130       max
#> 461  411.000 3.630       max
#> 240  412.000 1.130       max
#> 455  414.000 5.030       max
#> 47   415.000 2.970       max
#> 306  416.052 2.986       max
#> 27   416.667 4.250       max
#> 354  418.000 1.540       max
#> 40   422.000 2.260       max
#> 311  422.656 3.194       max
#> 279  424.000 1.840       max
#> 456  429.000 5.010       max
#> 310  429.768 3.571       max
#> 39   430.000 3.860       max
#> 7    431.000 3.600       max
#> 43   432.000 3.420       max
#> 460  435.000 3.720       max
#> 116  437.000 1.623       max
#> 117  438.000 1.676       max
#> 255  439.000 1.020       max
#> 347  441.000 3.780       max
#> 266  442.000 1.950       max
#> 378  444.000 3.550       max
#> 2    445.000 5.390       max
#> 224  450.000 0.800       max
#> 346  456.000 2.310       max
#> 236  458.000 1.170       max
#> 119  461.000 1.988       max
#> 249  462.000 1.840       max
#> 343  465.000 4.970       max
#> 342  470.000 2.990       max
#> 3    471.000 3.040       max
#> 349  489.000 2.330       max
#> 245  490.000 1.510       max
#> 348  503.000 1.430       max
#> 272  517.000 1.980       max
#> 258  520.000 1.980       max
#> 210  521.000 3.430       max
#> 212  522.000 3.800       max
#> 213  524.000 3.610       max
#> 276  525.000 2.160       max
#> 267  531.000 2.200       max
#> 209  536.000 3.260       max
#> 285  540.000 2.180       max
#> 263  543.000 1.960       max
#> 211  549.000 3.480       max
#> 254  552.000 2.200       max
#> 281  581.000 1.870       max
#> 2811 581.000 1.870       min
#> 2541 552.000 2.200       min
#> 2111 549.000 3.480       min
#> 2631 543.000 1.960       min
#> 2851 540.000 2.180       min
#> 2091 536.000 3.260       min
#> 2671 531.000 2.200       min
#> 2761 525.000 2.160       min
#> 2131 524.000 3.610       min
#> 2121 522.000 3.800       min
#> 2101 521.000 3.430       min
#> 2581 520.000 1.980       min
#> 2721 517.000 1.980       min
#> 3481 503.000 1.430       min
#> 2451 490.000 1.510       min
#> 3491 489.000 2.330       min
#> 37   471.000 3.040       min
#> 3421 470.000 2.990       min
#> 3431 465.000 4.970       min
#> 2491 462.000 1.840       min
#> 1191 461.000 1.988       min
#> 2361 458.000 1.170       min
#> 3461 456.000 2.310       min
#> 2241 450.000 0.800       min
#> 220  445.000 5.390       min
#> 118  444.000 1.886       min
#> 2661 442.000 1.950       min
#> 3471 441.000 3.780       min
#> 2551 439.000 1.020       min
#> 1171 438.000 1.676       min
#> 1161 437.000 1.623       min
#> 373  435.000 2.520       min
#> 4310 432.000 3.420       min
#> 710  431.000 3.600       min
#> 3910 430.000 3.860       min
#> 3101 429.768 3.571       min
#> 4561 429.000 5.010       min
#> 56   424.000 1.540       min
#> 3111 422.656 3.194       min
#> 374  422.000 2.020       min
#> 3541 418.000 1.540       min
#> 2710 416.667 4.250       min
#> 3061 416.052 2.986       min
#> 351  415.000 1.380       min
#> 4    414.000 3.100       min
#> 2401 412.000 1.130       min
#> 51   411.000 3.140       min
#> 2731 410.000 1.130       min
#> 3081 409.956 2.656       min
#> 191  409.524 4.300       min
#> 225  408.000 0.730       min
#> 2151 407.000 2.270       min
#> 2571 406.000 1.090       min
#> 2610 404.040 4.000       min
#> 365  404.000 2.050       min
#> 44   402.000 1.880       min
#> 55   401.000 2.800       min
#> 2410 400.952 4.210       min
#> 226  400.000 0.840       min
#> 138  399.000 2.542       min
#> 3261 398.684 2.502       min
#> 252  398.000 1.610       min
#> 375  396.000 1.280       min
#> 358  395.000 2.340       min
#> 218  393.000 1.480       min
#> 1211 392.000 1.373       min
#> 3251 391.537 2.374       min
#> 1710 391.176 3.990       min
#> 282  391.000 1.260       min
#> 11   390.000 3.270       min
#> 2510 389.744 4.560       min
#> 340  388.000 2.860       min
#> 120  387.000 1.155       min
#> 4711 386.000 3.870       min
#> 1231 385.000 1.589       min
#> 1810 384.536 3.730       min
#> 2681 382.000 1.620       min
#> 3271 381.984 2.609       min
#> 217  381.000 1.510       min
#> 221  380.000 2.190       min
#> 4661 379.000 2.420       min
#> 344  377.000 1.560       min
#> 284  376.000 2.060       min
#> 3041 375.412 3.833       min
#> 286  375.000 1.570       min
#> 891  374.194 3.480       min
#> 2751 373.000 1.600       min
#> 2771 372.000 1.540       min
#> 5601 371.925 5.589       min
#> 4651 371.000 2.650       min
#> 259  370.000 1.580       min
#> 4001 369.658 2.300       min
#> 2001 369.000 2.160       min
#> 2281 368.000 0.380       min
#> 2231 367.000 1.990       min
#> 208  366.000 1.850       min
#> 237  365.000 1.210       min
#> 1610 364.474 2.770       min
#> 246  363.000 1.240       min
#> 115  362.000 2.791       min
#> 881  361.053 3.430       min
#> 227  361.000 0.280       min
#> 1121 359.000 1.853       min
#> 1081 358.000 2.225       min
#> 377  357.000 2.050       min
#> 1581 356.250 2.850       min
#> 250  356.000 1.310       min
#> 3071 355.854 3.638       min
#> 2341 355.000 1.420       min
#> 1091 354.000 2.433       min
#> 1611 352.222 3.170       min
#> 3121 352.044 3.289       min
#> 339  352.000 2.190       min
#> 219  351.000 1.910       min
#> 2641 350.000 1.260       min
#> 3021 348.488 3.087       min
#> 356  348.000 2.760       min
#> 1801 347.000 3.440       min
#> 4671 346.000 1.940       min
#> 6251 344.806 4.338       min
#> 1771 344.000 3.440       min
#> 5581 343.679 4.513       min
#> 1441 343.333 2.060       min
#> 2041 343.000 1.460       min
#> 1821 342.000 3.390       min
#> 468  341.000 2.330       min
#> 207  340.000 1.850       min
#> 183  339.000 3.440       min
#> 4691 338.000 3.670       min
#> 125  336.000 1.750       min
#> 6641 335.316 3.321       min
#> 128  335.000 1.167       min
#> 370  334.000 2.860       min
#> 2911 333.844 2.397       min
#> 6691 333.161 2.959       min
#> 124  333.000 1.488       min
#> 3941 331.809 2.200       min
#> 206  331.000 1.470       min
#> 1641 330.769 4.300       min
#> 2891 330.713 2.692       min
#> 1311 330.000 1.618       min
#> 3001 329.946 2.872       min
#> 901  329.412 2.800       min
#> 130  329.000 1.520       min
#> 911  328.261 3.020       min
#> 62   328.000 2.700       min
#> 2901 327.894 2.833       min
#> 129  327.000 1.376       min
#> 1651 326.875 5.230       min
#> 3681 326.000 2.550       min
#> 3331 325.000 4.120       min
#> 4261 323.353 2.100       min
#> 4971 323.000 3.270       min
#> 2651 322.000 1.930       min
#> 5671 320.944 4.746       min
#> 159  320.000 1.920       min
#> 241  317.000 1.170       min
#> 1331 316.000 3.224       min
#> 3171 315.347 1.970       min
#> 199  315.000 1.990       min
#> 3711 314.000 2.080       min
#> 363  313.000 2.330       min
#> 4121 312.950 2.100       min
#> 6671 312.140 2.589       min
#> 106  312.000 1.992       min
#> 3951 311.892 1.600       min
#> 5571 311.579 5.070       min
#> 1041 311.000 1.759       min
#> 3161 310.184 1.782       min
#> 107  310.000 1.944       min
#> 238  309.000 1.980       min
#> 6621 307.093 3.242       min
#> 2441 307.000 1.190       min
#> 5591 306.502 4.519       min
#> 105  306.000 1.993       min
#> 6341 305.991 1.919       min
#> 6241 305.258 3.634       min
#> 4811 305.000 3.580       min
#> 5551 304.888 5.372       min
#> 5691 304.555 5.234       min
#> 5701 304.164 4.360       min
#> 188  304.000 3.110       min
#> 6861 303.518 5.644       min
#> 4981 303.000 3.200       min
#> 6331 302.636 2.397       min
#> 3211 302.249 3.665       min
#> 3921 302.245 2.000       min
#> 6291 302.059 2.991       min
#> 1411 301.667 1.810       min
#> 6851 301.220 5.806       min
#> 271  301.000 1.250       min
#> 2881 300.987 2.134       min
#> 6701 300.953 2.663       min
#> 6651 300.940 3.077       min
#> 5621 300.810 4.551       min
#> 2031 300.000 2.570       min
#> 3181 299.287 2.044       min
#> 2991 299.212 2.818       min
#> 2561 298.000 2.460       min
#> 6741 297.950 3.272       min
#> 3141 297.434 2.145       min
#> 2801 297.000 1.270       min
#> 6361 296.106 2.111       min
#> 64   296.000 2.400       min
#> 6661 295.141 3.264       min
#> 379  295.000 1.590       min
#> 5681 294.759 4.291       min
#> 3151 294.386 3.517       min
#> 6711 293.922 3.010       min
#> 5561 293.307 4.791       min
#> 247  293.000 1.850       min
#> 5651 292.992 6.065       min
#> 1521 292.500 2.340       min
#> 50   292.000 0.740       min
#> 3981 291.135 1.900       min
#> 69   291.000 1.700       min
#> 6301 290.987 3.315       min
#> 5631 290.752 3.794       min
#> 6261 290.345 2.555       min
#> 1401 290.000 1.160       min
#> 3091 289.814 3.436       min
#> 6471 289.700 3.131       min
#> 6151 289.048 2.980       min
#> 54   289.000 0.990       min
#> 6311 288.744 3.174       min
#> 5661 288.525 4.203       min
#> 1531 288.182 3.170       min
#> 4251 287.973 3.400       min
#> 2981 287.782 4.008       min
#> 2961 287.660 2.914       min
#> 5971 287.354 2.220       min
#> 3191 287.341 3.188       min
#> 6721 287.051 1.776       min
#> 3201 286.738 3.416       min
#> 2951 286.415 2.804       min
#> 5541 285.943 4.852       min
#> 59   285.000 1.500       min
#> 2871 284.555 1.824       min
#> 6771 284.446 3.721       min
#> 70   284.000 1.700       min
#> 6371 283.742 1.795       min
#> 6761 283.209 2.598       min
#> 3591 283.000 1.320       min
#> 6321 282.924 3.963       min
#> 5641 282.554 4.694       min
#> 2931 282.015 2.603       min
#> 239  282.000 0.000       min
#> 5611 281.540 4.470       min
#> 2971 281.432 2.367       min
#> 6681 281.378 2.891       min
#> 5521 281.270 4.458       min
#> 46   281.000 0.800       min
#> 3031 280.924 2.596       min
#> 5811 280.543 2.015       min
#> 6391 280.221 3.069       min
#> 4151 279.160 1.400       min
#> 6841 278.486 4.609       min
#> 6611 278.293 2.771       min
#> 335  278.000 0.960       min
#> 6091 277.125 2.450       min
#> 38   277.000 0.950       min
#> 5431 276.871 3.973       min
#> 42   276.000 0.810       min
#> 1721 275.714 3.860       min
#> 58   275.000 0.890       min
#> 6181 274.652 2.461       min
#> 2941 274.311 3.086       min
#> 6171 273.375 2.779       min
#> 5531 273.266 4.717       min
#> 6351 273.153 1.636       min
#> 6141 271.716 2.614       min
#> 5921 271.708 2.358       min
#> 341  271.000 1.130       min
#> 921  270.833 0.650       min
#> 4451 270.420 1.300       min
#> 4131 270.302 2.100       min
#> 6911 269.746 4.755       min
#> 5501 269.607 3.110       min
#> 5451 269.122 2.667       min
#> 68   269.000 1.600       min
#> 3321 268.000 1.340       min
#> 910  267.000 2.493       min
#> 6231 266.741 3.999       min
#> 5931 266.631 2.777       min
#> 5711 266.594 3.581       min
#> 3911 266.355 2.000       min
#> 57   266.000 2.430       min
#> 6131 265.293 2.045       min
#> 5961 264.824 2.474       min
#> 5491 264.751 3.030       min
#> 6811 264.057 2.953       min
#> 5731 263.937 2.229       min
#> 4411 263.650 1.500       min
#> 5441 263.581 2.763       min
#> 3051 263.398 2.219       min
#> 6801 263.293 2.631       min
#> 6401 263.021 2.690       min
#> 1951 263.000 2.400       min
#> 6081 262.445 1.914       min
#> 4901 262.000 2.400       min
#> 6381 261.681 2.327       min
#> 5951 260.873 1.880       min
#> 6201 260.686 3.135       min
#> 3891 260.487 1.700       min
#> 6781 260.076 3.670       min
#> 501  260.000 1.950       min
#> 5511 258.962 3.307       min
#> 6411 258.874 2.122       min
#> 6881 258.123 3.421       min
#> 2351 258.000 1.260       min
#> 6831 257.969 4.970       min
#> 6121 257.721 2.540       min
#> 6281 256.956 3.134       min
#> 6211 256.744 2.832       min
#> 6731 256.611 3.486       min
#> 3231 256.532 2.044       min
#> 84   256.000 1.638       min
#> 3811 255.546 0.600       min
#> 5471 255.216 2.709       min
#> 493  255.000 2.270       min
#> 5461 254.939 3.129       min
#> 3110 254.783 2.930       min
#> 3241 254.000 2.065       min
#> 4291 253.921 1.900       min
#> 851  253.571 2.130       min
#> 3861 252.373 0.800       min
#> 5841 252.197 3.468       min
#> 5421 251.704 2.858       min
#> 4051 251.548 2.200       min
#> 6871 251.300 5.614       min
#> 3221 251.229 1.829       min
#> 6101 251.011 1.597       min
#> 3011 250.952 3.289       min
#> 6041 250.559 2.391       min
#> 4071 250.274 2.100       min
#> 5981 250.231 2.056       min
#> 5061 250.000 2.680       min
#> 4271 249.968 4.700       min
#> 3901 249.330 1.400       min
#> 6011 249.135 2.050       min
#> 6451 247.799 1.699       min
#> 5821 247.713 2.403       min
#> 5481 247.045 3.779       min
#> 355  247.000 0.830       min
#> 4021 246.302 2.100       min
#> 253  246.000 1.290       min
#> 5991 245.947 2.001       min
#> 6631 245.713 3.502       min
#> 3721 245.000 1.550       min
#> 329  244.000 1.420       min
#> 3961 243.874 1.200       min
#> 6791 243.530 2.929       min
#> 4531 243.310 2.400       min
#> 507  243.000 2.570       min
#> 4061 242.430 3.200       min
#> 248  242.000 0.480       min
#> 2691 241.000 1.090       min
#> 6221 240.809 3.895       min
#> 4161 240.735 1.900       min
#> 6191 240.594 2.778       min
#> 6441 240.173 2.181       min
#> 4921 240.000 2.130       min
#> 4111 239.684 1.300       min
#> 4381 239.620 3.800       min
#> 4091 239.055 1.900       min
#> 5081 238.000 2.310       min
#> 5041 237.000 2.680       min
#> 5831 236.518 3.012       min
#> 5871 236.408 3.656       min
#> 5101 236.000 2.320       min
#> 1671 235.833 2.830       min
#> 5941 235.606 2.543       min
#> 3931 235.533 1.400       min
#> 1511 235.000 0.470       min
#> 4081 234.004 1.600       min
#> 251  234.000 1.010       min
#> 6161 233.607 2.647       min
#> 5851 233.474 3.139       min
#> 6571 233.059 3.092       min
#> 5021 233.000 1.550       min
#> 6751 232.798 3.707       min
#> 6821 232.768 4.013       min
#> 3841 231.986 1.300       min
#> 931  230.435 1.060       min
#> 137  230.000 0.690       min
#> 6271 229.636 2.963       min
#> 5861 229.303 2.364       min
#> 1661 229.167 2.750       min
#> 3451 229.000 1.340       min
#> 6001 228.925 2.202       min
#> 6531 228.449 2.947       min
#> 4231 228.401 0.800       min
#> 5771 228.264 2.224       min
#> 951  227.027 0.840       min
#> 5291 227.000 2.020       min
#> 4431 226.990 2.900       min
#> 4461 226.760 1.500       min
#> 6601 226.430 2.585       min
#> 5751 226.040 2.307       min
#> 260  226.000 1.080       min
#> 6901 225.965 4.564       min
#> 4321 225.087 0.000       min
#> 6491 223.954 1.836       min
#> 5801 223.794 1.813       min
#> 6561 223.777 4.694       min
#> 6421 223.715 2.169       min
#> 4371 223.380 3.400       min
#> 5881 223.029 2.537       min
#> 5401 223.000 3.840       min
#> 2921 222.862 1.433       min
#> 4201 222.062 1.800       min
#> 6511 221.884 2.107       min
#> 6541 221.239 2.946       min
#> 6521 221.026 3.072       min
#> 6891 220.688 3.772       min
#> 283  220.339 2.600       min
#> 3131 220.218 1.170       min
#> 361  220.000 0.790       min
#> 4481 219.900 1.600       min
#> 971  218.947 2.080       min
#> 3851 218.847 2.900       min
#> 5781 218.788 2.810       min
#> 5791 218.624 2.025       min
#> 6461 218.428 1.304       min
#> 4041 218.055 2.000       min
#> 6501 217.229 1.375       min
#> 1461 216.667 0.650       min
#> 5891 216.316 2.881       min
#> 4341 216.250 3.300       min
#> 6431 216.162 1.441       min
#> 751  216.000 3.062       min
#> 941  215.152 0.710       min
#> 524  215.000 2.080       min
#> 5721 214.733 2.035       min
#> 6551 214.563 3.119       min
#> 5391 214.000 1.950       min
#> 4421 213.970 2.600       min
#> 505  213.000 1.710       min
#> 5761 212.946 2.532       min
#> 5901 212.128 3.580       min
#> 6591 211.544 3.522       min
#> 4491 211.390 2.200       min
#> 5111 211.000 1.730       min
#> 150  210.000 0.630       min
#> 4011 209.551 2.000       min
#> 4281 209.027 2.100       min
#> 2210 207.937 2.620       min
#> 3971 207.195 1.900       min
#> 741  207.000 2.823       min
#> 6071 206.334 1.851       min
#> 4941 206.000 1.900       min
#> 6061 205.466 1.728       min
#> 1481 205.000 0.410       min
#> 6031 204.835 3.169       min
#> 5911 204.174 3.201       min
#> 5411 204.000 3.300       min
#> 328  203.000 0.810       min
#> 6111 202.923 1.794       min
#> 1361 202.500 0.810       min
#> 2781 202.000 0.890       min
#> 2910 201.408 2.860       min
#> 5741 201.344 2.028       min
#> 4141 201.173 1.000       min
#> 262  200.000 1.240       min
#> 6581 199.591 3.348       min
#> 2421 199.000 0.180       min
#> 1471 197.500 0.790       min
#> 2331 197.000 0.170       min
#> 5131 196.000 2.070       min
#> 4441 195.700 2.600       min
#> 3010 195.683 2.720       min
#> 4361 195.660 2.300       min
#> 4881 195.000 3.210       min
#> 4101 194.031 1.600       min
#> 4241 192.752 1.600       min
#> 4351 192.310 1.800       min
#> 1571 191.667 1.150       min
#> 72   191.000 1.646       min
#> 1621 190.833 2.290       min
#> 4471 190.390 1.600       min
#> 80   189.000 1.114       min
#> 3991 188.440 1.900       min
#> 4871 188.000 2.840       min
#> 1421 186.667 1.120       min
#> 961  186.567 1.250       min
#> 5301 186.000 0.940       min
#> 6021 185.796 2.333       min
#> 4501 185.720 1.900       min
#> 1431 185.714 1.300       min
#> 4961 185.000 0.600       min
#> 4391 184.810 1.300       min
#> 4031 184.232 1.100       min
#> 761  183.000 2.832       min
#> 6481 182.858 2.342       min
#> 4521 181.190 2.100       min
#> 871  181.046 2.770       min
#> 4331 180.170 1.300       min
#> 330  180.000 0.900       min
#> 3881 179.462 1.100       min
#> 5141 178.000 2.010       min
#> 216  177.698 2.470       min
#> 1701 177.500 1.420       min
#> 4401 176.100 0.800       min
#> 861  175.472 0.930       min
#> 5311 175.000 0.870       min
#> 1711 174.286 1.220       min
#> 4191 174.062 1.000       min
#> 4841 173.000 2.600       min
#> 4171 171.065 1.200       min
#> 771  167.000 1.554       min
#> 1491 166.667 0.500       min
#> 991  166.071 0.930       min
#> 202  164.773 1.450       min
#> 2311 164.000 1.680       min
#> 4311 162.437 1.400       min
#> 1341 160.000 0.320       min
#> 229  159.000 1.490       min
#> 1681 158.333 0.950       min
#> 2310 158.156 2.230       min
#> 135  155.000 0.310       min
#> 6051 153.828 2.103       min
#> 2301 153.000 1.390       min
#> 3601 152.000 0.740       min
#> 357  151.000 0.380       min
#> 5321 150.000 0.440       min
#> 5221 146.000 1.250       min
#> 3871 144.678 0.800       min
#> 5361 144.000 1.660       min
#> 1691 142.857 1.000       min
#> 5341 141.000 2.090       min
#> 981  139.474 0.530       min
#> 3821 139.074 1.100       min
#> 5231 139.000 0.960       min
#> 4511 137.150 1.900       min
#> 3831 135.703 1.000       min
#> 5351 133.000 1.450       min
#> 4221 132.095 1.000       min
#> 4211 129.827 1.400       min
#> 4301 127.445 1.500       min
#> 1551 123.750 0.990       min
#> 1541 120.000 0.960       min
#> 5331 116.000 1.640       min
#> 5381 114.000 1.070       min
#> 4181 108.619 0.800       min
#>