BLA: Boundary Line Analysis

Fits boundary line models to datasets as proposed by Webb (1972) doi:10.1080/00221589.1972.11514472 and makes statistical inferences about their parameters. Provides additional tools for testing datasets for evidence of boundary presence and selecting initial starting values for model optimization prior to fitting the boundary line models. It also includes tools for conducting post-hoc analyses such as predicting boundary values and identifying the most limiting factor (Miti, Milne, Giller, Lark (2024) doi:10.1016/j.fcr.2024.109365 ). This ensures a comprehensive analysis for datasets that exhibit upper boundary structures.

References

Lark, R. M., & Milne, A. E. (2016). Boundary line analysis of the effect of water filled pore space on nitrous oxide emission from cores of arable soil. European Journal of Soil Science, 67 , 148-159.

Milne, A. E., Ferguson, R. B., & Lark, R. M. (2006). Estimating a boundary line model for a biological response by maximum likelihood. Annals of Applied Biology, 149, 223–234.

Schnug, E., Heym, J. M., & Murphy, D. P. L. (1995). Boundary line determination technique (BOLIDES). In P. C. Robert, R. H. Rust, & W. E. Larson (Eds.), site specific management for agricultural systems (p. 899-908). Wiley Online Library.

Webb, R. A. (1972). Use of the boundary line in analysis of biological data. Journal of Horticultural Science, 47, 309-319.

See also

Author

Maintainer: Chawezi Miti chawezi.miti@nottingham.ac.uk (ORCID) [copyright holder]

Authors:

Richard M Lark (ORCID)

Alice E Milne (ORCID)

Ken E Giller (ORCID)

Other contributors:

Victor O Sadras [contributor]

University of Nottingham/Rothamsted Research [funder]