Determination of the most limiting factor to biological response

This function determines the most limiting factor based on von Liebig law of the minimum given results of the predicted boundary line values for the different factors of interest. Boundary lines for various factors are fitted and the factor that predicts the minimum response for a particular point is considered as the most limiting factor (Casanova et al. 1995).

Usage

limfactor(...)

Arguments

...

vectors with predicted values from the boundary line models for each factor being evaluated.

Value

A dataframe consisting of the most limiting factor and the minimum predicted response

Author

Chawezi Miti <chawezi.miti@nottingham.ac.uk>

Examples

N<-rnorm(10,50,5) #assuming these are predicted responses using the fitted BL for N,P,K
K<-rnorm(10,50,4)
P<-rnorm(10,50,6)

limfactor(N,K,P)
#> [[1]]
#>          Rs Lim_factor
#> 1  44.51065          N
#> 2  44.58357          P
#> 3  40.97562          P
#> 4  36.15900          P
#> 5  45.35519          P
#> 6  47.36602          K
#> 7  43.22327          K
#> 8  45.07431          K
#> 9  55.01983          K
#> 10 43.35657          P
#> 
#> [[2]]
#> [1] 63.16171
#>